pyproject.toml
[tool.poetry]
name = "crag"
version = "0.1.0"
description = "Cognitive Graph-RAG: An Agentic Framework for Dynamic Reasoning on Knowledge Graphs"
authors = ["Surya P", "Vidyullata", "T. Sai Mayithri", "Swastik Mohapatra"]
readme = "README.md"
packages = [{include = "crag", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
networkx = "^3.2.1"
requests = "^2.31.0"
numpy = "^1.26.0"
tqdm = "^4.66.0"
torch = "^2.2.0"
transformers = "^4.38.0"
# SPARQLWrapper = "^2.0.0" # Optional, can use requests
faiss-cpu = "^1.8.0" # For vector store
matplotlib = "^3.8.0" # For visualization
pandas = "^2.2.0" # For experiment manager
scikit-learn = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
